changed cast from long to int for printing
git-svn-id: http://internalbrr.cr.usgs.gov/svn_MOWS/PRMS_SVN/mmf/trunk@6975 6316d359-22eb-4c0b-9c0f-3bdc8ec1e293

+++ write_vstats.c
@@ -64,8 +64,9 @@ void write_vstats (FILE *statvar_file) {
       break;
 
     case M_LONG:
-
-      (void)fprintf(statvar_file,"%ld ", *(long *)stat_list->value);
+	// markstro test
+    //  (void)fprintf(statvar_file,"%ld ", *(long *)stat_list->value);
+	(void)fprintf(statvar_file,"%ld ", *(int *)stat_list->value);
       break;
 
     }
